backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0
pydantic>=1.8.0
tensorflow>=2.13.0
numpy>=1.21.0
pillow>=8.3.0
python-multipart>=0.0.5
aiofiles>=0.7.0
